+++ src/Services/StateFilterService.php
@@ -6,8 +6,10 @@ use Filament\Tables\Filters\SelectFilter;
 
 class StateFilterService extends _BaseService
 {
-    public function tableFilter(string $field = 'state', bool $multiple = true, ?string $label = null): SelectFilter
+    public function tableFilter(?string $field = null, bool $multiple = true, ?string $label = null): SelectFilter
     {
+        $field = $field ?? $this->field;
+
         return SelectFilter::make($field)
             ->label($label ?? __($field))
             ->options($this->getOptions())
@@ -19,9 +21,9 @@ class StateFilterService extends _BaseService
     {
         $result = [];
 
-        foreach ($this->getStates() ?? [] as $option) {
+        foreach ($this->getStates() as $option) {
             if ($option::includeToFilters() || in_array($option, $this->include_states)) {
-                $result[$option] = $option::title();
+                $result[$option::getMorphClass()] = $option::title();
             }
         }